requirements.txt
requests~=2.25.1
pip~=20.3
PyYAML~=5.3.1
certifi~=2020.12.5
chardet~=3.0.4
urllib3~=1.26.2
six~=1.15.0
idna~=2.10
setuptools~=49.2.1
docutils~=0.15.2
python-dateutil~=2.8.1